+++ packages/EasyAsync/bundle/config/messenger_stop_on_messages_limit.php
@@ -1,9 +1,10 @@
 <?php
 declare(strict_types=1);
 
+namespace Symfony\Component\DependencyInjection\Loader\Configurator;
+
 use EonX\EasyAsync\Bundle\Enum\ConfigParam;
 use EonX\EasyAsync\Messenger\Subscriber\StopWorkerOnMessagesLimitSubscriber;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
@@ -13,7 +14,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
 
     $services
         ->set(StopWorkerOnMessagesLimitSubscriber::class)
-        ->arg('$minMessages', '%' . ConfigParam::MessengerWorkerStopMinMessages->value . '%')
-        ->arg('$maxMessages', '%' . ConfigParam::MessengerWorkerStopMaxMessages->value . '%')
+        ->arg('$minMessages', param(ConfigParam::MessengerWorkerStopMinMessages->value))
+        ->arg('$maxMessages', param(ConfigParam::MessengerWorkerStopMaxMessages->value))
         ->tag('monolog.logger', ['channel' => ConfigParam::LogChannel->value]);
 };
