+++ packages/EasyAsync/composer.json
@@ -31,7 +31,9 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "EonX\\EasyAsync\\Tests\\": "tests"
+            "EonX\\EasyAsync\\Tests\\": "tests",
+            "EonX\\EasyAsync\\Tests\\Unit\\": "tests/Unit/src",
+            "EonX\\EasyAsync\\Tests\\Unit\\Bundle\\": "tests/Unit/bundle"
         }
     },
     "suggest": {
